retool-for-life/backend/requirements.txt
openai>=1.0.0
openai-agents>=0.0.17
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
twilio>=8.10.0
httpx>=0.26.0
python-dotenv>=1.0.0
pandas>=2.1.0
websockets>=12.0
python-multipart>=0.0.6
aiofiles>=23.2.1
requests>=2.31.0
nest-asyncio>=1.5.8